overall the best way for current situation is
- to generate certs first in CI host -> run_tests.sh: used by both unit and integration tests
- generate in gradle only if needed, e.g. local ./gradlew clean test
  - we will not clean certs with gradle clean, if we do integration tests will fail due to mismatch
  - and genera.sh will skip if certs exist - makes sure we are not generating twice

+++ utils/ssl-certs-generator/generate.sh
@@ -5,6 +5,11 @@ cd "$(dirname "$0")"
 # This is for testing purposes only
 set -e
 
+if [[ -f generated/root.crt ]]; then
+    echo "Certificates already exist. Skipping..."
+    exit 0
+fi
+
 rm -rf generated
 mkdir generated
 cd generated
